requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
djangorestframework==3.15.1
fonttools==4.55.0
imbalanced-learn==0.12.4
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.2
statsmodels==0.14.4
threadpoolctl==3.5.0
tzdata==2024.2
xgboost==2.1.2
